CIR-1193-thread-error-handling
Exceptions raised in threads were not being bubbled up to the caller because `Thread.current.abort_on_exception` was not set to true.
When this is the case and threads are communicating via `IO.pipe`, the calling thread may hang indefinitely if the other thread has died without closing its end of the pipe.

- Set `Thread.current.abort_on_exception = true` for child threads.
- Move writing to pipe to main thread, and reading from pipe to child thread.
- Handle exceptions, cleaning up pipes and threads appropriately.